requirements.txt
flask==2.3.3
pycryptodome==3.19.0
requests==2.31.0 